# src/main/java/com/consort/database/entities/NewProject.java
package com.consort.database.entities;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import com.fasterxml.jackson.annotation.JsonProperty;

@JsonIgnoreProperties
public class NewProject {
    @JsonProperty("name")
    private String name;

    @JsonProperty("description")
    private String description;

    public NewProject() {}

    public String getName() {
        return name;
    }

    public void setName(final String name) {
        this.name = name;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(final String description) {
        this.description = description;
    }
}